requirements.txt
pandas
numpy
matplotlib
ipython
ipykernel
tqdm
pickle5
coverage
dependency_injector
pyyaml